Backend/requirements.txt
fastapi==0.116.2
uvicorn==0.36.0
pydantic==2.11.9
python-dotenv==1.1.1
pymongo==4.15.1
langchain==0.3.27
langchain-google-genai==2.1.12
qdrant-client==1.15.1
loguru==0.7.3
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]
langchain-text-splitters
beautifulsoup4==4.12.3
lxml==5.3.0
playwright
requests==2.32.3
sentence_transformers==5.1.1
pydantic[email]
# Testing dependencies
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.24.1